+++ packages/auth-construct/src/construct.ts
@@ -107,6 +107,8 @@ export class AmplifyAuth
 
   private readonly oAuthSettings: cognito.OAuthSettings | undefined;
 
+  private readonly name: string;
+
   /**
    * Create a new Auth construct with AuthProps.
    * If no props are provided, email login and defaults will be used.
@@ -118,11 +120,13 @@ export class AmplifyAuth
   ) {
     super(scope, id);
 
+    this.name = props.name ?? '';
+
     // UserPool
     this.computedUserPoolProps = this.getUserPoolProps(props);
     this.userPool = new cognito.UserPool(
       this,
-      'UserPool',
+      `${this.name}UserPool`,
       this.computedUserPoolProps
     );
 
@@ -137,7 +141,7 @@ export class AmplifyAuth
       this.domainPrefix &&
       this.providerSetupResult.providersList.length > 0
     ) {
-      this.userPool.addDomain('UserPoolDomain', {
+      this.userPool.addDomain(`${this.name}UserPoolDomain`, {
         cognitoDomain: { domainPrefix: this.domainPrefix },
       });
     } else if (
@@ -178,7 +182,7 @@ export class AmplifyAuth
     // UserPool Client
     const userPoolClient = new cognito.UserPoolClient(
       this,
-      'UserPoolAppClient',
+      `${this.name}UserPoolAppClient`,
       {
         userPool: this.userPool,
         authFlows: DEFAULTS.AUTH_FLOWS,
@@ -253,7 +257,7 @@ export class AmplifyAuth
    */
   private setupAuthAndUnAuthRoles = (identityPoolId: string): DefaultRoles => {
     const result: DefaultRoles = {
-      auth: new Role(this, 'authenticatedUserRole', {
+      auth: new Role(this, `${this.name}authenticatedUserRole`, {
         assumedBy: new FederatedPrincipal(
           'cognito-identity.amazonaws.com',
           {
@@ -267,7 +271,7 @@ export class AmplifyAuth
           'sts:AssumeRoleWithWebIdentity'
         ),
       }),
-      unAuth: new Role(this, 'unauthenticatedUserRole', {
+      unAuth: new Role(this, `${this.name}unauthenticatedUserRole`, {
         assumedBy: new FederatedPrincipal(
           'cognito-identity.amazonaws.com',
           {
@@ -295,14 +299,19 @@ export class AmplifyAuth
   ) => {
     // setup identity pool
     const region = Stack.of(this).region;
-    const identityPool = new cognito.CfnIdentityPool(this, 'IdentityPool', {
-      allowUnauthenticatedIdentities: DEFAULTS.ALLOW_UNAUTHENTICATED_IDENTITIES,
-    });
+    const identityPool = new cognito.CfnIdentityPool(
+      this,
+      `${this.name}IdentityPool`,
+      {
+        allowUnauthenticatedIdentities:
+          DEFAULTS.ALLOW_UNAUTHENTICATED_IDENTITIES,
+      }
+    );
     const roles = this.setupAuthAndUnAuthRoles(identityPool.ref);
     const identityPoolRoleAttachment =
       new cognito.CfnIdentityPoolRoleAttachment(
         this,
-        'IdentityPoolRoleAttachment',
+        `${this.name}IdentityPoolRoleAttachment`,
         {
           identityPoolId: identityPool.ref,
           roles: {
@@ -611,7 +620,7 @@ export class AmplifyAuth
       const googleProps = external.google;
       result.google = new cognito.UserPoolIdentityProviderGoogle(
         this,
-        'GoogleIdP',
+        `${this.name}GoogleIdP`,
         {
           userPool,
           clientId: googleProps.clientId,
@@ -633,7 +642,7 @@ export class AmplifyAuth
     if (external.facebook) {
       result.facebook = new cognito.UserPoolIdentityProviderFacebook(
         this,
-        'FacebookIDP',
+        `${this.name}FacebookIDP`,
         {
           userPool,
           ...external.facebook,
@@ -654,7 +663,7 @@ export class AmplifyAuth
     if (external.loginWithAmazon) {
       result.amazon = new cognito.UserPoolIdentityProviderAmazon(
         this,
-        'AmazonIDP',
+        `${this.name}AmazonIDP`,
         {
           userPool,
           ...external.loginWithAmazon,
@@ -676,7 +685,7 @@ export class AmplifyAuth
     if (external.signInWithApple) {
       result.apple = new cognito.UserPoolIdentityProviderApple(
         this,
-        'AppleIDP',
+        `${this.name}AppleIDP`,
         {
           userPool,
           ...external.signInWithApple,
@@ -696,36 +705,44 @@ export class AmplifyAuth
       result.providersList.push('APPLE');
     }
     if (external.oidc) {
-      result.oidc = new cognito.UserPoolIdentityProviderOidc(this, 'OidcIDP', {
-        userPool,
-        ...external.oidc,
-        attributeMapping:
-          external.oidc.attributeMapping ?? shouldMapEmailAttributes
-            ? {
-                email: {
-                  attributeName: 'email',
-                },
-              }
-            : undefined,
-      });
+      result.oidc = new cognito.UserPoolIdentityProviderOidc(
+        this,
+        `${this.name}OidcIDP`,
+        {
+          userPool,
+          ...external.oidc,
+          attributeMapping:
+            external.oidc.attributeMapping ?? shouldMapEmailAttributes
+              ? {
+                  email: {
+                    attributeName: 'email',
+                  },
+                }
+              : undefined,
+        }
+      );
       result.providersList.push('OIDC');
     }
     if (external.saml) {
       const saml = external.saml;
-      result.saml = new cognito.UserPoolIdentityProviderSaml(this, 'SamlIDP', {
-        userPool,
-        attributeMapping: saml.attributeMapping,
-        identifiers: saml.identifiers,
-        idpSignout: saml.idpSignout,
-        metadata: {
-          metadataContent: saml.metadata.metadataContent,
-          metadataType:
-            saml.metadata.metadataType === 'FILE'
-              ? UserPoolIdentityProviderSamlMetadataType.FILE
-              : UserPoolIdentityProviderSamlMetadataType.URL,
-        },
-        name: saml.name,
-      });
+      result.saml = new cognito.UserPoolIdentityProviderSaml(
+        this,
+        `${this.name}SamlIDP`,
+        {
+          userPool,
+          attributeMapping: saml.attributeMapping,
+          identifiers: saml.identifiers,
+          idpSignout: saml.idpSignout,
+          metadata: {
+            metadataContent: saml.metadata.metadataContent,
+            metadataType:
+              saml.metadata.metadataType === 'FILE'
+                ? UserPoolIdentityProviderSamlMetadataType.FILE
+                : UserPoolIdentityProviderSamlMetadataType.URL,
+          },
+          name: saml.name,
+        }
+      );
       result.providersList.push('SAML');
     }
     return result;
@@ -894,7 +911,7 @@ export class AmplifyAuth
         output.oauthRedirectSignOut = this.oAuthSettings.logoutUrls
           ? this.oAuthSettings.logoutUrls.join(',')
           : '';
-        output.webClientId = this.resources.userPoolClient.userPoolClientId;
+        output.oauthClientId = this.resources.userPoolClient.userPoolClientId;
         output.oauthResponseType = 'code';
       }
     }