+++ packages/cli/src/commands/pipeline-deploy/pipeline_deploy_command_factory.ts
@@ -1,5 +1,6 @@
 import { CommandModule } from 'yargs';
 import { BackendDeployerFactory } from '@aws-amplify/backend-deployer';
+import { PackageManagerControllerFactory } from '@aws-amplify/cli-core';
 
 import {
   PipelineDeployCommand,
@@ -7,6 +8,7 @@ import {
 } from './pipeline_deploy_command.js';
 import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
 import { ClientConfigGeneratorAdapter } from '../../client-config/client_config_generator_adapter.js';
+import { printer } from '../../printer.js';
 
 /**
  * Creates pipeline deploy command
@@ -19,6 +21,13 @@ export const createPipelineDeployCommand = (): CommandModule<
   const clientConfigGenerator = new ClientConfigGeneratorAdapter(
     credentialProvider
   );
-  const backendDeployer = BackendDeployerFactory.getInstance();
+  const packageManagerControllerFactory = new PackageManagerControllerFactory(
+    process.cwd(),
+    printer
+  );
+  const backendDeployerFactory = new BackendDeployerFactory(
+    packageManagerControllerFactory.getPackageManagerController()
+  );
+  const backendDeployer = backendDeployerFactory.getInstance();
   return new PipelineDeployCommand(clientConfigGenerator, backendDeployer);
 };