+++ packages/create-amplify/src/amplify_project_creator.ts
@@ -1,9 +1,8 @@
 import { LogLevel } from '@aws-amplify/cli-core';
-import { PackageManagerController } from './package_manager_controller.js';
+import { PackageManagerController } from '@aws-amplify/plugin-types';
 import { ProjectRootValidator } from './project_root_validator.js';
-import { InitialProjectFileGenerator } from './initial_project_file_generator.js';
-import { NpmProjectInitializer } from './npm_project_initializer.js';
 import { GitIgnoreInitializer } from './gitignore_initializer.js';
+import { InitialProjectFileGenerator } from './initial_project_file_generator.js';
 import { printer } from './printer.js';
 
 const LEARN_MORE_USAGE_DATA_TRACKING_LINK = `https://docs.amplify.aws/gen2/reference/telemetry`;
@@ -28,12 +27,11 @@ export class AmplifyProjectCreator {
    * Delegates out to other classes that handle parts of the getting started experience
    */
   constructor(
+    private readonly projectRoot: string,
     private readonly packageManagerController: PackageManagerController,
     private readonly projectRootValidator: ProjectRootValidator,
-    private readonly initialProjectFileGenerator: InitialProjectFileGenerator,
-    private readonly npmInitializedEnsurer: NpmProjectInitializer,
     private readonly gitIgnoreInitializer: GitIgnoreInitializer,
-    private readonly projectRoot: string
+    private readonly initialProjectFileGenerator: InitialProjectFileGenerator
   ) {}
 
   /**
@@ -46,7 +44,7 @@ export class AmplifyProjectCreator {
     );
     await this.projectRootValidator.validate();
 
-    await this.npmInitializedEnsurer.ensureInitialized();
+    await this.packageManagerController.initializeProject();
 
     await printer.indicateProgress(
       `Installing required dependencies`,
@@ -71,15 +69,18 @@ export class AmplifyProjectCreator {
 
     printer.log('Successfully created a new project!');
 
-    const cdCommand =
+    const cdPreamble =
       process.cwd() === this.projectRoot
         ? ''
-        : `cd .${this.projectRoot.replace(process.cwd(), '')}; `;
+        : `Navigate to your project directory using
+'cd .${this.projectRoot.replace(process.cwd(), '')}'.
+Then get started with the following commands:
+`;
 
     printer.log(
       `Welcome to AWS Amplify! 
-Run \`npx amplify help\` for a list of available commands. 
-Get started by running \`${cdCommand}npx amplify sandbox\`.`
+${cdPreamble}
+${this.packageManagerController.getWelcomeMessage()}`
     );
 
     printer.log(