+++ packages/plugin-types/src/index.ts
@@ -15,3 +15,4 @@ export * from './backend_identifier.js';
 export * from './deployment_type.js';
 export * from './resource_access_acceptor.js';
 export * from './ssm_environment_entries_generator.js';
+export * from './package_manager_controller.js';